requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
reportlab>=4.0.0
openpyxl>=3.1.0
shap>=0.42.0
Pillow>=10.0.0
